refactor: move config package from internal to public
Exposes config types and Init() so external consumers (e.g. keyline-operator)
can import github.com/The127/Keyline/config directly.
