+++ cmd/api/main.go
@@ -12,7 +12,7 @@ package main
 import (
 	"github.com/The127/Keyline/internal/authentication"
 	"github.com/The127/Keyline/internal/commands"
-	"github.com/The127/Keyline/internal/config"
+	"github.com/The127/Keyline/config"
 	"github.com/The127/Keyline/internal/database"
 	"github.com/The127/Keyline/internal/jobs"
 	"github.com/The127/Keyline/internal/logging"
