+++ internal/events/usercreated.go
@@ -1,7 +1,7 @@
 package events
 
 import (
-	"github.com/The127/Keyline/internal/config"
+	"github.com/The127/Keyline/config"
 	db "github.com/The127/Keyline/internal/database"
 	"github.com/The127/Keyline/internal/messages"
 	"github.com/The127/Keyline/internal/middlewares"
