+++ internal/handlers/oidc.go
@@ -2,7 +2,7 @@ package handlers
 
 import (
 	"github.com/The127/Keyline/api"
-	"github.com/The127/Keyline/internal/config"
+	"github.com/The127/Keyline/config"
 	"github.com/The127/Keyline/internal/database"
 	"github.com/The127/Keyline/internal/jsonTypes"
 	"github.com/The127/Keyline/internal/middlewares"
