+++ internal/handlers/virtualServers.go
@@ -3,7 +3,7 @@ package handlers
 import (
 	"github.com/The127/Keyline/api"
 	"github.com/The127/Keyline/internal/commands"
-	"github.com/The127/Keyline/internal/config"
+	"github.com/The127/Keyline/config"
 	"github.com/The127/Keyline/internal/middlewares"
 	"github.com/The127/Keyline/internal/queries"
 	"github.com/The127/Keyline/utils"
