+++ internal/middlewares/SessionMiddleware.go
@@ -1,7 +1,7 @@
 package middlewares
 
 import (
-	"github.com/The127/Keyline/internal/config"
+	"github.com/The127/Keyline/config"
 	"github.com/The127/Keyline/utils"
 	"context"
 	"errors"
