+++ internal/services/keyValue/kvStore.go
@@ -1,7 +1,7 @@
 package keyValue
 
 import (
-	"github.com/The127/Keyline/internal/config"
+	"github.com/The127/Keyline/config"
 	"github.com/The127/Keyline/internal/middlewares"
 	"context"
 	"errors"
