+++ tests/e2e/harness.go
@@ -4,7 +4,7 @@ import (
 	"github.com/The127/Keyline/client"
 	"github.com/The127/Keyline/internal/authentication"
 	"github.com/The127/Keyline/internal/commands"
-	"github.com/The127/Keyline/internal/config"
+	"github.com/The127/Keyline/config"
 	"github.com/The127/Keyline/internal/database"
 	"github.com/The127/Keyline/internal/database/postgres"
 	"github.com/The127/Keyline/internal/middlewares"
