+++ tests/integration/role_flow_test.go
@@ -2,7 +2,7 @@ package integration
 
 import (
 	"github.com/The127/Keyline/internal/commands"
-	"github.com/The127/Keyline/internal/config"
+	"github.com/The127/Keyline/config"
 	"github.com/The127/Keyline/internal/queries"
 	"github.com/The127/Keyline/utils"
